VendingMachine
Write an interface specfication for a model class that represents a vending machine. In exchange for money, the machine distributes coffee or tea. When you design the specification, consider the machine's attributes first: the quantities of coffee, tea, and excess change the machine holds. (Don't forget to remember the amount of money a customer has inserted!) When you design the machine's methods, consider the machine's responsibilities to give coffee, tea, and change, to refund a customer's coins when asked, and to refuse service if no coffee or tea is left.